requirements.txt
Pillow==9.5.0
PyPDF2==3.0.1
PyYAML==6.0
reportlab==3.6.12